+++ solar/solarsystemcamera.h
@@ -12,7 +12,7 @@ class SolarSystemCameraMan : public CameraMan
 public:
     SolarSystemCameraMan(Scene &scene, Viewport viewport);
 
-    void update(const Inputs& inputs, double dt) override;
+    void update(const Inputs& inputs, float dt) override;
 
     void handleKeyboard(const Inputs& inputs, const KeyboardEvent& event) override;
     void handleMouseMove(const Inputs& inputs, const MouseMoveEvent& event) override;
@@ -22,7 +22,7 @@ public:
 
     void setMode(Mode mode);
     void setBodyIndex(int bodyId);
-    void setDistance(double dist);
+    void setDistance(float dist);
 
     void zoomIn();
     void zoomOut();
@@ -40,10 +40,10 @@ public:
     void strafeUp();
     void strafeDown();
 
-    static const double ZOON_INC;
-    static const double EXPOSURE_INC;
-    static const double ROTATE_INC;
-    static const double APPROACH_INC;
+    static const float ZOON_INC;
+    static const float EXPOSURE_INC;
+    static const float ROTATE_INC;
+    static const float APPROACH_INC;
 
 protected:
     void orbit(int newBodyId, int oldBodyId);
